Create Fibonacci
Using a Loop: The fibonacci_loop function initializes a list with the first two Fibonacci numbers (0 and 1). It then uses a while loop to generate the rest of the sequence by appending the sum of the last two numbers until the desired length is reached.
Using Recursion: The fibonacci_recursive function calls itself with decremented values to calculate the nth Fibonacci number. The fibonacci_sequence_recursive function generates the entire sequence by calling fibonacci_recursive for each index up to n.